Support for WIZnet W5100 ethernet module SPI_EN
The W5100 module has a SPI_EN signal that enables or disables the SPI
bus.
A bug in the module causes MISO to be continously driven by the module
when SPI_EN is pulled high. This is described in this application note:
https://www.google.se/url?sa=t&rct=j&q=&esrc=s&source=web&cd=1&cad=rja&uact=8&ved=0CCAQFjAA&url=http%3A%2F%2Fwww.wiznet.co.kr%2FUpLoad_Files%2FReferenceFiles%2FW5100_AN_SPI.pdf&ei=ij84VK7uM6nNygPqvICYDw&usg=AFQjCNGz1qttBEwG1azc1Eza4j69v1rpDQ&sig2=UUUtb4Dr0F35g8ByHMsyig&bvm=bv.77161500,d.bGQ

This commit changes the Ethernet gateway sketch to drive SPI_EN high or
low, depending on which module is in use (SPI bus is shared between
Ethernet and RF modules).
The pin attached to SPI_EN is defined by W5100_SPI_EN. This define can
be commented out if SPI_EN is not needed (different module) or if the
module manages SPI_EN in hardware (some W5100 shields does this).