NRF5 Platform support
This is the support of the Nordic Semiconductor NRF51 (nrf51822,
nrf51422) and NRF52 (nrf52832) platform. Based on arduino-nRF5
(https://github.com/sandeepmistry/arduino-nRF5).

The Radio is working in an NRF24 compatible mode. BLE is not
supported.

The included "NVM" driver allows an EEPROM like access to the
last 16/32k of Flash pages. It can easily ported to ESP8266 or
SAMD by providing a Flash interface.