Save memory by eliminating sha256_hash array
eliminated sha256_hash array to save memory. Instead pass the
Sha256Class result pointer.

There seems to be no use for the extra sha256_hash array as it was
zeroed exactly at the same time, the Sha256Class result was invalidated
(by calling signerSha256Init

If the result needs to stay around while another sha is calculated,
it's better to have the user taking care of this.