Generate Doxygen from configure (#952)
* Generate Doxygen from configure

Use case: view the ./configure help text on the web.
This has been requested in
https://forum.mysensors.org/post/77751

Manually putting the help text in the forum or on mysensors.org
would lead to diverging documentation, no version management and
extra maintenance burden. This commit attempts to auto-generate
the documentation and use existing Doxygen infrastructure to
add the help text to the rest of the MySensors documentation.