Add missing keywords
A blacklist is available at the end of keywords.txt

I decided to blacklist all the LINUX keywords. I guess people
could use the Arduino IDE to edit the sketch for the raspberry
pi gateway, but I'm not sure we should encourage that.

I'll add gitler support for auto-detecting missing keywords
in a separate PR.