Remove redundant keyword identifiers (#1140)
The third field of keywords.txt is used to provide Arduino Language
Reference links, which are accessed from the Arduino IDE by
highlighting the keyword and then selecting "Find in Reference"
from the Help or right click menu. Adding values to this field that
do not match any existing reference pages results in a "Could not
open the URL" error.