Remove unnecessary MY_IP_RENEWAL_INTERVAL* (#1429)
* Remove unnecessary MY_IP_RENEWAL_INTERVAL*

Ethernet.maintain already checks when we need to renew, based on
the lease time given by the DHCP server.

https://www.arduino.cc/en/Reference/EthernetMaintain recommends
"just call it on every loop() invocation" so there is no
need for us to keep a separate timer.

* Remove call to _w5100_spi_en

This function doesn't enable SPI, so it should not
disable it either. Let calling functions handle SPI
enable/disable.